split tests for main and AutoRelease
